[flutter_tools] remove global usage of flutterUsage from gradle (#76330)
