+++ packages/flutter_tools/lib/src/android/gradle.dart
@@ -22,7 +22,7 @@ import '../build_info.dart';
 import '../cache.dart';
 import '../convert.dart';
 import '../flutter_manifest.dart';
-import '../globals.dart' as globals hide logger, printStatus, printTrace, printError, processManager, processUtils, fs, artifacts;
+import '../globals.dart' as globals hide logger, printStatus, printTrace, printError, processManager, processUtils, fs, artifacts, flutterUsage;
 import '../project.dart';
 import '../reporting/reporting.dart';
 import 'android_builder.dart';
@@ -129,14 +129,6 @@ bool _isSupportedVersion(AndroidProject project) {
   return false;
 }
 
-/// Returns the apk file created by [buildGradleProject]
-Future<File> getGradleAppOut(AndroidProject androidProject) async {
-  if (!_isSupportedVersion(androidProject)) {
-    _exitWithUnsupportedProjectMessage();
-  }
-  return getApkDirectory(androidProject.parent).childFile('app.apk');
-}
-
 /// Runs `gradlew dependencies`, ensuring that dependencies are resolved and
 /// potentially downloaded.
 Future<void> checkGradleDependencies(Logger logger, ProcessUtils processUtils) async {
@@ -217,15 +209,18 @@ class AndroidGradleBuilder implements AndroidBuilder {
     @required ProcessManager processManager,
     @required FileSystem fileSystem,
     @required Artifacts artifacts,
+    @required Usage usage,
   }) : _logger = logger,
        _fileSystem = fileSystem,
        _artifacts = artifacts,
+       _usage = usage,
        _processUtils = ProcessUtils(logger: logger, processManager: processManager);
 
   final Logger _logger;
   final ProcessUtils _processUtils;
   final FileSystem _fileSystem;
   final Artifacts _artifacts;
+  final Usage _usage;
 
   /// Builds the AAR and POM files for the current Flutter module or plugin.
   @override
@@ -336,18 +331,18 @@ class AndroidGradleBuilder implements AndroidBuilder {
     assert(globals.androidSdk != null);
 
     if (!project.android.isUsingGradle) {
-      _exitWithProjectNotUsingGradleMessage();
+      _exitWithProjectNotUsingGradleMessage(_usage);
     }
     if (!_isSupportedVersion(project.android)) {
-      _exitWithUnsupportedProjectMessage();
+      _exitWithUnsupportedProjectMessage(_usage);
     }
     final Directory buildDirectory = project.android.buildDirectory;
 
     final bool usesAndroidX = isAppUsingAndroidX(project.android.hostAppGradleRoot);
     if (usesAndroidX) {
-      BuildEvent('app-using-android-x', flutterUsage: globals.flutterUsage).send();
+      BuildEvent('app-using-android-x', flutterUsage: _usage).send();
     } else if (!usesAndroidX) {
-      BuildEvent('app-not-using-android-x', flutterUsage: globals.flutterUsage).send();
+      BuildEvent('app-not-using-android-x', flutterUsage: _usage).send();
       _logger.printStatus("$warningMark Your app isn't using AndroidX.", emphasis: true);
       _logger.printStatus(
         'To avoid potential build failures, you can quickly migrate your app '
@@ -488,11 +483,11 @@ class AndroidGradleBuilder implements AndroidBuilder {
       status.stop();
     }
 
-    globals.flutterUsage.sendTiming('build', 'gradle', sw.elapsed);
+    _usage.sendTiming('build', 'gradle', sw.elapsed);
 
     if (exitCode != 0) {
       if (detectedGradleError == null) {
-        BuildEvent('gradle-unknown-failure', flutterUsage: globals.flutterUsage).send();
+        BuildEvent('gradle-unknown-failure', flutterUsage: _usage).send();
         throwToolExit(
           'Gradle task $assembleTask failed with exit code $exitCode',
           exitCode: exitCode,
@@ -518,7 +513,7 @@ class AndroidGradleBuilder implements AndroidBuilder {
                 shouldBuildPluginAsAar: shouldBuildPluginAsAar,
                 retries: retries - 1,
               );
-              BuildEvent(successEventLabel, flutterUsage: globals.flutterUsage).send();
+              BuildEvent(successEventLabel, flutterUsage: _usage).send();
               return;
             case GradleBuildStatus.retryWithAarPlugins:
               await buildGradleApp(
@@ -530,13 +525,13 @@ class AndroidGradleBuilder implements AndroidBuilder {
                 shouldBuildPluginAsAar: true,
                 retries: retries - 1,
               );
-              BuildEvent(successEventLabel, flutterUsage: globals.flutterUsage).send();
+              BuildEvent(successEventLabel, flutterUsage: _usage).send();
               return;
             case GradleBuildStatus.exit:
             // noop.
           }
         }
-        BuildEvent('gradle-${detectedGradleError.eventLabel}-failure', flutterUsage: globals.flutterUsage).send();
+        BuildEvent('gradle-${detectedGradleError.eventLabel}-failure', flutterUsage: _usage).send();
         throwToolExit(
           'Gradle task $assembleTask failed with exit code $exitCode',
           exitCode: exitCode,
@@ -545,7 +540,7 @@ class AndroidGradleBuilder implements AndroidBuilder {
     }
 
     if (isBuildingBundle) {
-      final File bundleFile = findBundleFile(project, buildInfo, _logger);
+      final File bundleFile = findBundleFile(project, buildInfo, _logger, _usage);
       final String appSize = (buildInfo.mode == BuildMode.debug)
           ? '' // Don't display the size when building a debug variant.
           : ' (${getSizeAsMB(bundleFile.lengthSync())})';
@@ -562,7 +557,7 @@ class AndroidGradleBuilder implements AndroidBuilder {
     }
     // Gradle produced an APK.
     final Iterable<String> apkFilesPaths = project.isModule
-        ? findApkFilesModule(project, androidBuildInfo, _logger)
+        ? findApkFilesModule(project, androidBuildInfo, _logger, _usage)
         : listApkPaths(androidBuildInfo);
     final Directory apkDirectory = getApkDirectory(project);
     final File apkFile = apkDirectory.childFile(apkFilesPaths.first);
@@ -571,6 +566,7 @@ class AndroidGradleBuilder implements AndroidBuilder {
         project: project,
         fileExtension: '.apk',
         logger: _logger,
+        usage: _usage,
       );
     }
 
@@ -603,7 +599,7 @@ class AndroidGradleBuilder implements AndroidBuilder {
     final SizeAnalyzer sizeAnalyzer = SizeAnalyzer(
       fileSystem: _fileSystem,
       logger: _logger,
-      flutterUsage: globals.flutterUsage,
+      flutterUsage: _usage,
     );
     final String archName = getNameForAndroidArch(androidBuildInfo.targetArchs.single);
     final BuildInfo buildInfo = androidBuildInfo.buildInfo;
@@ -760,7 +756,7 @@ class AndroidGradleBuilder implements AndroidBuilder {
     } finally {
       status.stop();
     }
-    globals.flutterUsage.sendTiming('build', 'gradle-aar', sw.elapsed);
+    _usage.sendTiming('build', 'gradle-aar', sw.elapsed);
 
     if (result.exitCode != 0) {
       _logger.printStatus(result.stdout, wrap: false);
@@ -830,12 +826,11 @@ class AndroidGradleBuilder implements AndroidBuilder {
       } on ToolExit {
         // Log the entire plugin entry in `.flutter-plugins` since it
         // includes the plugin name and the version.
-        BuildEvent('gradle-plugin-aar-failure', eventError: plugin, flutterUsage: globals.flutterUsage).send();
+        BuildEvent('gradle-plugin-aar-failure', eventError: plugin, flutterUsage: _usage).send();
         throwToolExit('The plugin $pluginName could not be built due to the issue above.');
       }
     }
   }
-
 }
 
 /// Prints how to consume the AAR from a host app.
@@ -911,8 +906,8 @@ String _calculateSha(File file) {
   return _hex(sha1.convert(bytes).bytes);
 }
 
-void _exitWithUnsupportedProjectMessage() {
-  BuildEvent('unsupported-project', eventError: 'gradle-plugin', flutterUsage: globals.flutterUsage).send();
+void _exitWithUnsupportedProjectMessage(Usage usage) {
+  BuildEvent('unsupported-project', eventError: 'gradle-plugin', flutterUsage: usage).send();
   throwToolExit(
     '$warningMark Your app is using an unsupported Gradle project. '
     'To fix this problem, create a new project by running `flutter create -t app <app-directory>` '
@@ -920,14 +915,22 @@ void _exitWithUnsupportedProjectMessage() {
   );
 }
 
-void _exitWithProjectNotUsingGradleMessage() {
-  BuildEvent('unsupported-project', eventError: 'app-not-using-gradle', flutterUsage: globals.flutterUsage).send();
+void _exitWithProjectNotUsingGradleMessage(Usage usage) {
+  BuildEvent('unsupported-project', eventError: 'app-not-using-gradle', flutterUsage: usage).send();
   throwToolExit(
     '$warningMark The build process for Android has changed, and the '
     'current project configuration is no longer valid. Please consult\n\n'
     'https://github.com/flutter/flutter/wiki/Upgrading-Flutter-projects-to-build-with-gradle\n\n'
     'for details on how to upgrade the project.'
   );
+}
+
+/// Returns the apk file created by [buildGradleProject]
+Future<File> getGradleAppOut(AndroidProject androidProject, Usage usage) async {
+  if (!_isSupportedVersion(androidProject)) {
+    _exitWithUnsupportedProjectMessage(usage);
+  }
+  return getApkDirectory(androidProject.parent).childFile('app.apk');
 }
 
 /// Returns [true] if the current app uses AndroidX.
@@ -947,6 +950,7 @@ Iterable<String> findApkFilesModule(
   FlutterProject project,
   AndroidBuildInfo androidBuildInfo,
   Logger logger,
+  Usage usage,
 ) {
   final Iterable<String> apkFileNames = _apkFilesFor(androidBuildInfo);
   final Directory apkDirectory = getApkDirectory(project);
@@ -980,6 +984,7 @@ Iterable<String> findApkFilesModule(
       project: project,
       fileExtension: '.apk',
       logger: logger,
+      usage: usage,
     );
   }
   return apks.map((File file) => file.path);
@@ -1018,7 +1023,7 @@ Iterable<String> listApkPaths(
 }
 
 @visibleForTesting
-File findBundleFile(FlutterProject project, BuildInfo buildInfo, Logger logger) {
+File findBundleFile(FlutterProject project, BuildInfo buildInfo, Logger logger, Usage usage) {
   final List<File> fileCandidates = <File>[
     getBundleDirectory(project)
       .childDirectory(camelCase(buildInfo.modeName))
@@ -1053,6 +1058,7 @@ File findBundleFile(FlutterProject project, BuildInfo buildInfo, Logger logger) 
     project: project,
     fileExtension: '.aab',
     logger: logger,
+    usage: usage,
   );
   return null;
 }
@@ -1062,6 +1068,7 @@ void _exitWithExpectedFileNotFound({
   @required FlutterProject project,
   @required String fileExtension,
   @required Logger logger,
+  @required Usage usage,
 }) {
   assert(project != null);
   assert(fileExtension != null);
@@ -1072,7 +1079,7 @@ void _exitWithExpectedFileNotFound({
     settings:
     'androidGradlePluginVersion: $androidGradlePluginVersion, '
       'fileExtension: $fileExtension',
-    flutterUsage: globals.flutterUsage,
+    flutterUsage: usage,
   ).send();
   throwToolExit(
     'Gradle build failed to produce an $fileExtension file. '
